# 8278518: String(byte[], int, int, Charset) constructor and String.translateEscapes() miss bounds check elimination