@AlanBateman 2021-12-13
> Originally this was spotted by by Amir Hadadi in https://stackoverflow.com/questions/70272651/missing-bounds-checking-elimination-in-string-constructor

Before anyone looks at this, can you confirm that the patch does not include any code or tests/benchmarks that were taken from SO?