@amirhadadi 2021-12-14
@AlanBateman @stsypanov I have no problem signing the OCA.
However like @stsypanov mentioned, I don't think this is the right approach.
Adding explicit check that the offset is non negative makes no semantic difference, it just helps the optimizer to figure out that no bounds checking is needed.
A better approach is to improve the optimizer so that it can apply bounds checking elimination in this case (and hopefully in many other cases).