@merykitty 2021-12-14
Hi, the problem here does not seem to lie in the bound check eliminations, as the patched version still contains the bound checks in these lines:

    13.70%            │  0x00007f6b88eb6079:   cmp    0x8(%rsp),%r10d
     0.01%            │  0x00007f6b88eb607e:   jae    0x00007f6b88eb6671

The problem, at first glance, seems to be that our compiled code is trying to compute this mysterious number

    long temp = sl;
    loop {
        long newTemp = (long)((int)((temp - (long)offset + 1) & (-2L)) + offset) - 2L;
        if (newTemp u>= temp) jump;
    }

I think further investigation is needed, maybe our compiler is trying too hard to deal with various appearance of `offset` in the loop.