@stsypanov 2021-12-14
> The problem, at first glance, seems to be that our compiled code is trying to compute this mysterious number

@merykitty how do you view it? 🤔