@stsypanov 2021-12-15
@AlanBateman I suggest to decide first whether this should be fixed on HotSpot level (which is the best option to me) or on Java level.

If we choose HotSpot then the issue should be reassigned, because I cannot do the fix myself.

If we choose Java then I would revert the change for String constructor and Amir would commit it into this branch having OCA signed.

What do you think?