@dean-long 2021-12-17
This does look like a HotSpot JIT compiler issue to me.  My guess is that it is related to how checkBoundsOffCount() checks for offset < 0:

396          if ((length | fromIndex | size) < 0 || size > length - fromIndex)

using | to combine three values.