@amirhadadi 2021-12-18
@dean-long actually the issue reproduces with Java 17 where `checkBoundsOffCount` was implemented in a more straight forward way:

```
    static void checkBoundsOffCount(int offset, int count, int length) {
        if (offset < 0 || count < 0 || offset > length - count) {
            throw new StringIndexOutOfBoundsException(
                "offset " + offset + ", count " + count + ", length " + length);
        }
    }

```

Here's a [gist](https://gist.github.com/amirhadadi/9505c3f5d9ad68cad2fbfd1b9e01f0b8) with a benchmark you can run. In this benchmark I'm comparing safe and unsafe reads from the byte array  (I didn't modify the code to add the offset >= 0 condition).

Here are the results:

```
Benchmark                       Mode  Cnt    Score    Error  Units
StringBenchmark.safeDecoding    avgt   20  120.312 ± 11.674  ns/op
StringBenchmark.unsafeDecoding  avgt   20   72.628 ±  0.479  ns/op

```