@amirhadadi 2021-12-21
@rwestrel thanks for the correction!

Here are the updated results:

```
Benchmark                       Mode  Cnt    Score   Error  Units
StringBenchmark.safeDecoding    avgt   20  113.849 ± 1.609  ns/op
StringBenchmark.unsafeDecoding  avgt   20   85.272 ± 1.462  ns/op

```