@stsypanov 2022-01-12
As soon as the proper fix is provided in https://github.com/openjdk/jdk/pull/7034 I close this PR.